backend/app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Mongo stack (compatible)
motor==3.7.1
pymongo>=4.9,<5

# Auth & security
passlib==1.7.4
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
email-validator==2.2.0

# HTTP client for DigitalOcean Agent
httpx==0.27.2
